requirements.txt
PySimpleGUI~=4.60.1
winsdk==1.0.0b5; platform_system == "Windows"
PyGObject~=3.42.2; platform_system == "Linux"
dbus-python~=1.2.18; platform_system == "Linux"
python-benedict~=0.25.1
requests~=2.28.1
python-telegram-bot~=13.13
pyqt5~=5.15.7